requirements.txt
setuptools==58.1.0
pydantic>=1.9.0
paho-mqtt>=1.6.1
tenacity>=8.1.0